refactor dashboard to make room for video and lighting tools in the future
