+++ src/pages/AllStagePlots.tsx
@@ -371,11 +371,11 @@ const AllStagePlots = () => {
       <main className="flex-grow container mx-auto px-4 py-12 mt-12">
         <div className="flex items-center justify-between mb-8">
           <button
-            onClick={() => navigate("/dashboard")}
+            onClick={() => navigate("/audio")}
             className="flex items-center text-gray-400 hover:text-white transition-colors"
           >
             <ArrowLeft className="h-5 w-5 mr-2" />
-            Back to Dashboard
+            Back to Audio
           </button>
 
           <button
