+++ src/pages/ProductionPage.tsx
@@ -1,0 +1,660 @@
+import React, { useEffect, useState, useRef } from "react";
+import { useNavigate, Link } from "react-router-dom";
+import { supabase } from "../lib/supabase";
+import Header from "../components/Header";
+import Footer from "../components/Footer";
+import {
+  PlusCircle,
+  Trash2,
+  Edit,
+  Download,
+  List,
+  AlertTriangle,
+  CalendarDays,
+  ClipboardList,
+  Loader,
+  Info,
+  ArrowLeftCircle,
+} from "lucide-react";
+import html2canvas from "html2canvas";
+import ExportModal from "../components/ExportModal";
+import ProductionScheduleExport from "../components/production-schedule/ProductionScheduleExport";
+import PrintProductionScheduleExport from "../components/production-schedule/PrintProductionScheduleExport";
+import RunOfShowExport from "../components/run-of-show/RunOfShowExport";
+import PrintRunOfShowExport from "../components/run-of-show/PrintRunOfShowExport";
+import { ScheduleForExport, DetailedScheduleItem } from "./ProductionScheduleEditor";
+import { LaborScheduleItem } from "../components/production-schedule/ProductionScheduleLabor";
+import { RunOfShowItem, CustomColumnDefinition } from "./RunOfShowEditor";
+import { v4 as uuidv4 } from 'uuid';
+
+interface ProductionScheduleSummary {
+  id: string;
+  name: string;
+  created_at: string;
+  last_edited?: string;
+}
+
+interface FullProductionScheduleData {
+  id: string;
+  name: string;
+  created_at: string;
+  last_edited?: string;
+  user_id: string;
+  show_name?: string;
+  job_number?: string;
+  facility_name?: string;
+  project_manager?: string;
+  production_manager?: string;
+  account_manager?: string;
+  set_datetime?: string;
+  strike_datetime?: string;
+  crew_key?: Array<{ id: string; name: string; color: string }>;
+  detailed_schedule_items?: DetailedScheduleItem[];
+  labor_schedule_items?: LaborScheduleItem[];
+}
+
+interface RunOfShowSummary {
+  id: string;
+  name: string;
+  created_at: string;
+  last_edited?: string;
+}
+
+export interface FullRunOfShowData {
+  id: string;
+  name: string;
+  created_at: string;
+  last_edited?: string;
+  user_id: string;
+  items: RunOfShowItem[];
+  custom_column_definitions: CustomColumnDefinition[];
+}
+
+const parseDateTime = (dateTimeStr: string | null | undefined) => {
+  if (!dateTimeStr) return { date: undefined, time: undefined, full: undefined };
+  try {
+    const d = new Date(dateTimeStr);
+    if (isNaN(d.getTime())) return { date: dateTimeStr, time: undefined, full: dateTimeStr };
+    return {
+      date: d.toISOString().split('T')[0],
+      time: d.toTimeString().split(' ')[0].substring(0, 5),
+      full: dateTimeStr,
+    };
+  } catch (e) {
+    return { date: dateTimeStr, time: undefined, full: dateTimeStr };
+  }
+};
+
+const transformToScheduleForExport = (fullSchedule: FullProductionScheduleData): ScheduleForExport => {
+  const setDateTimeParts = parseDateTime(fullSchedule.set_datetime);
+  const strikeDateTimeParts = parseDateTime(fullSchedule.strike_datetime);
+
+  return {
+    id: fullSchedule.id || uuidv4(),
+    name: fullSchedule.name,
+    created_at: fullSchedule.created_at || new Date().toISOString(),
+    last_edited: fullSchedule.last_edited,
+    info: {
+      event_name: fullSchedule.show_name,
+      job_number: fullSchedule.job_number,
+      venue: fullSchedule.facility_name,
+      project_manager: fullSchedule.project_manager,
+      production_manager: fullSchedule.production_manager,
+      account_manager: fullSchedule.account_manager,
+      date: setDateTimeParts.date,
+      load_in: setDateTimeParts.time,
+      event_start: setDateTimeParts.time,
+      event_end: strikeDateTimeParts.time,
+      strike_datetime: fullSchedule.strike_datetime,
+    },
+    crew_key: fullSchedule.crew_key?.map(ck => ({ ...ck })) || [],
+    detailed_schedule_items: fullSchedule.detailed_schedule_items?.map(item => ({
+      ...item,
+      assigned_crew_ids: item.assigned_crew_ids || (item.assigned_crew_id ? [item.assigned_crew_id] : [])
+    })) || [],
+    labor_schedule_items: fullSchedule.labor_schedule_items?.map(item => ({ ...item })) || [],
+  };
+};
+
+const ProductionPage = () => {
+  const navigate = useNavigate();
+  const [loading, setLoading] = useState(true);
+  const [user, setUser] = useState<any>(null);
+  const [productionSchedules, setProductionSchedules] = useState<ProductionScheduleSummary[]>([]);
+  const [runOfShows, setRunOfShows] = useState<RunOfShowSummary[]>([]);
+
+  const [exportingItemId, setExportingItemId] = useState<string | null>(null);
+
+  const [currentExportProductionSchedule, setCurrentExportProductionSchedule] = useState<ScheduleForExport | null>(null);
+  const [showProductionScheduleExportModal, setShowProductionScheduleExportModal] = useState(false);
+  const [exportProductionScheduleId, setExportProductionScheduleId] = useState<string | null>(null);
+
+  const [currentExportRunOfShow, setCurrentExportRunOfShow] = useState<FullRunOfShowData | null>(null);
+  const [showRunOfShowExportModal, setShowRunOfShowExportModal] = useState(false);
+  const [exportRunOfShowId, setExportRunOfShowId] = useState<string | null>(null);
+
+  const [showDeleteConfirm, setShowDeleteConfirm] = useState(false);
+  const [documentToDelete, setDocumentToDelete] = useState<{
+    id: string;
+    type: "schedule" | "runofshow";
+    name: string;
+  } | null>(null);
+  const [supabaseError, setSupabaseError] = useState<string | null>(null);
+
+  const productionScheduleExportRef = useRef<HTMLDivElement>(null);
+  const printProductionScheduleExportRef = useRef<HTMLDivElement>(null);
+  const runOfShowExportRef = useRef<HTMLDivElement>(null);
+  const printRunOfShowExportRef = useRef<HTMLDivElement>(null);
+
+  useEffect(() => {
+    const checkUser = async () => {
+      try {
+        const { data, error } = await supabase.auth.getUser();
+        if (error) throw error;
+
+        if (data.user) {
+          setUser(data.user);
+          await Promise.all([
+            fetchProductionSchedules(data.user.id),
+            fetchRunOfShows(data.user.id),
+          ]);
+        } else {
+          navigate("/login");
+        }
+      } catch (error) {
+        console.error("Error checking auth status:", error);
+        setSupabaseError("Failed to connect to Supabase. Please check your configuration.");
+        navigate("/login");
+      } finally {
+        setLoading(false);
+      }
+    };
+
+    checkUser();
+  }, [navigate]);
+  
+  const handleSignOut = async () => {
+    try {
+      await supabase.auth.signOut();
+      navigate("/");
+    } catch (error) {
+      console.error("Error signing out:", error);
+    }
+  };
+
+  const fetchProductionSchedules = async (userId: string) => {
+    try {
+      const { data, error } = await supabase
+        .from("production_schedules")
+        .select("id, name, created_at, last_edited")
+        .eq("user_id", userId)
+        .order("created_at", { ascending: false });
+      if (error) throw error;
+      if (data) setProductionSchedules(data as ProductionScheduleSummary[]);
+    } catch (error) {
+      console.error("Error fetching production schedules:", error);
+      setSupabaseError("Failed to fetch production schedules.");
+    }
+  };
+
+  const fetchRunOfShows = async (userId: string) => {
+    try {
+      const { data, error } = await supabase
+        .from("run_of_shows")
+        .select("id, name, created_at, last_edited")
+        .eq("user_id", userId)
+        .order("created_at", { ascending: false });
+      if (error) throw error;
+      if (data) setRunOfShows(data as RunOfShowSummary[]);
+    } catch (error) {
+      console.error("Error fetching run of shows:", error);
+      setSupabaseError("Failed to fetch run of shows.");
+    }
+  };
+
+  const handleCreateProductionSchedule = () => navigate("/production-schedule/new");
+  const handleCreateRunOfShow = () => navigate("/run-of-show/new");
+
+  const handleDeleteRequest = (id: string, type: "schedule" | "runofshow", name: string) => {
+    setDocumentToDelete({ id, type, name });
+    setShowDeleteConfirm(true);
+  };
+
+  const confirmDelete = async () => {
+    if (!documentToDelete) return;
+    try {
+      let tableName = "";
+      if (documentToDelete.type === "schedule") tableName = "production_schedules";
+      else if (documentToDelete.type === "runofshow") tableName = "run_of_shows";
+
+      if (tableName) {
+        const { error } = await supabase.from(tableName).delete().eq("id", documentToDelete.id);
+        if (error) throw error;
+
+        if (documentToDelete.type === "schedule") {
+          setProductionSchedules(productionSchedules.filter((item) => item.id !== documentToDelete.id));
+        } else if (documentToDelete.type === "runofshow") {
+          setRunOfShows(runOfShows.filter((item) => item.id !== documentToDelete.id));
+        }
+      }
+    } catch (error) {
+      console.error("Error deleting document:", error);
+      setSupabaseError("Failed to delete document. Please try again.");
+    } finally {
+      setShowDeleteConfirm(false);
+      setDocumentToDelete(null);
+    }
+  };
+
+  const cancelDelete = () => {
+    setShowDeleteConfirm(false);
+    setDocumentToDelete(null);
+  };
+
+  const handleEditProductionSchedule = (id: string) => navigate(`/production-schedule/${id}`);
+  const handleEditRunOfShow = (id: string) => navigate(`/run-of-show/${id}`);
+
+  const handleExportClick = (id: string, type: 'schedule' | 'runofshow') => {
+    if (type === 'schedule') {
+      setExportProductionScheduleId(id);
+      setShowProductionScheduleExportModal(true);
+    } else if (type === 'runofshow') {
+      setExportRunOfShowId(id);
+      setShowRunOfShowExportModal(true);
+    }
+  };
+
+  const exportImageWithCanvas = async (
+    targetRef: React.RefObject<HTMLDivElement>,
+    itemName: string,
+    fileNameSuffix: string,
+    backgroundColor: string,
+    font: string
+  ) => {
+    if (!targetRef.current) {
+      console.error("Export component ref not ready.");
+      setSupabaseError("Export component not ready. Please try again.");
+      return;
+    }
+    
+    await new Promise(resolve => setTimeout(resolve, 100));
+
+    if (document.fonts && typeof document.fonts.ready === 'function') {
+      try { await document.fonts.ready; } 
+      catch (fontError) { console.warn("Error waiting for document fonts to be ready:", fontError); }
+    } else {
+      await new Promise(resolve => setTimeout(resolve, 400));
+    }
+
+    try {
+      const canvas = await html2canvas(targetRef.current!, {
+        scale: 2, backgroundColor, useCORS: true, logging: process.env.NODE_ENV === 'development', letterRendering: true,
+        onclone: (clonedDoc) => {
+          const styleGlobal = clonedDoc.createElement('style');
+          styleGlobal.innerHTML = `* { font-family: ${font}, sans-serif !important; vertical-align: baseline !important; }`;
+          clonedDoc.head.appendChild(styleGlobal);
+          clonedDoc.body.style.fontFamily = `${font}, sans-serif`;
+          Array.from(clonedDoc.querySelectorAll('*')).forEach((el: any) => {
+            if (el.style) { el.style.fontFamily = `${font}, sans-serif`; el.style.verticalAlign = 'baseline';}
+          });
+        }
+      });
+      const image = canvas.toDataURL("image/png");
+      const link = document.createElement("a");
+      link.download = `${itemName.replace(/\s+/g, "-").toLowerCase()}-${fileNameSuffix}.png`;
+      link.href = image;
+      link.click();
+    } catch (error) {
+      console.error(`Error exporting ${fileNameSuffix}:`, error);
+      setSupabaseError(`Failed to export ${fileNameSuffix}. See console for details.`);
+    }
+  };
+
+  const prepareAndExecuteProductionScheduleExport = async (scheduleId: string, type: 'image' | 'print') => {
+    setExportingItemId(scheduleId);
+    setShowProductionScheduleExportModal(false);
+    try {
+      const { data: rawData, error } = await supabase.from("production_schedules").select("*").eq("id", scheduleId).single();
+      if (error || !rawData) throw error || new Error("Production schedule not found");
+      
+      const transformedData = transformToScheduleForExport(rawData as FullProductionScheduleData);
+      setCurrentExportProductionSchedule(transformedData);
+      await new Promise(resolve => setTimeout(resolve, 50));
+
+      if (type === 'image') {
+        await exportImageWithCanvas(productionScheduleExportRef, transformedData.name, "production-schedule", "#0f172a", "Inter");
+      } else {
+        await exportImageWithCanvas(printProductionScheduleExportRef, transformedData.name, "production-schedule-print", "#ffffff", "Arial");
+      }
+    } catch (err) {
+      console.error("Error preparing production schedule export:", err);
+      setSupabaseError("Failed to prepare production schedule for export.");
+    } finally {
+      setCurrentExportProductionSchedule(null);
+      setExportingItemId(null);
+      setExportProductionScheduleId(null);
+    }
+  };
+
+  const prepareAndExecuteRunOfShowExport = async (runOfShowId: string, type: 'image' | 'print') => {
+    setExportingItemId(runOfShowId);
+    setShowRunOfShowExportModal(false);
+    try {
+      const { data, error } = await supabase.from("run_of_shows").select("*").eq("id", runOfShowId).single();
+      if (error || !data) throw error || new Error("Run of Show not found");
+      
+      const fullData = {
+        ...data,
+        items: data.items || [],
+        custom_column_definitions: data.custom_column_definitions || [],
+      } as FullRunOfShowData;
+
+      setCurrentExportRunOfShow(fullData);
+      await new Promise(resolve => setTimeout(resolve, 50));
+
+      if (type === 'image') {
+        await exportImageWithCanvas(runOfShowExportRef, fullData.name, "run-of-show", "#0f172a", "Inter");
+      } else {
+        await exportImageWithCanvas(printRunOfShowExportRef, fullData.name, "run-of-show-print", "#ffffff", "Arial");
+      }
+    } catch (err) {
+      console.error("Error preparing Run of Show export:", err);
+      setSupabaseError("Failed to prepare Run of Show for export.");
+    } finally {
+      setCurrentExportRunOfShow(null);
+      setExportingItemId(null);
+      setExportRunOfShowId(null);
+    }
+  };
+
+  if (loading) {
+    return (
+      <div className="min-h-screen bg-gray-900 flex items-center justify-center">
+        <Loader className="animate-spin rounded-full h-12 w-12 border-t-2 border-b-2 border-indigo-500" />
+      </div>
+    );
+  }
+
+  return (
+    <div className="min-h-screen bg-gray-900 flex flex-col">
+      <Header onSignOut={handleSignOut} />
+
+      {supabaseError && (
+        <div className="bg-red-500 text-white px-4 py-3 shadow-sm">
+          <div className="container mx-auto flex items-center">
+            <Info className="h-5 w-5 mr-2" />
+            <div>
+              <p className="font-medium">Error</p>
+              <p className="text-sm">{supabaseError}</p>
+            </div>
+          </div>
+        </div>
+      )}
+
+      <main className="flex-grow container mx-auto px-4 py-12 mt-12">
+        <div className="mb-8">
+          <Link
+            to="/dashboard"
+            className="inline-flex items-center text-indigo-400 hover:text-indigo-300 transition-colors duration-200 mb-4 group"
+          >
+            <ArrowLeftCircle className="h-5 w-5 mr-2 transform group-hover:-translate-x-1 transition-transform duration-200" />
+            Back to Dashboard
+          </Link>
+          <h1 className="text-4xl font-bold text-white mb-4">Production Documents</h1>
+          <p className="text-lg text-gray-300">Manage your production schedules and run of shows.</p>
+        </div>
+
+        <div className="grid grid-cols-1 md:grid-cols-2 gap-8 mb-12">
+          {/* My Production Schedules Card */}
+          <div className="bg-gray-800 p-6 rounded-xl shadow-md hover:shadow-lg transition-all duration-200">
+            <div className="flex justify-between items-start mb-6">
+              <div>
+                <h2 className="text-xl font-semibold text-white mb-2">My Production Schedules</h2>
+                <p className="text-gray-400">Plan and track your event timelines</p>
+              </div>
+              <CalendarDays className="h-8 w-8 text-indigo-400" />
+            </div>
+            <div className="space-y-4">
+              {productionSchedules.length > 0 ? (
+                <div className="space-y-3">
+                  {productionSchedules.slice(0, 3).map((schedule) => (
+                    <div
+                      key={schedule.id}
+                      className="bg-gray-700 p-4 rounded-lg flex justify-between items-center"
+                    >
+                      <div>
+                        <h3 className="text-white font-medium">{schedule.name}</h3>
+                        <p className="text-gray-400 text-sm">
+                          {schedule.last_edited
+                            ? `Edited ${new Date(schedule.last_edited).toLocaleDateString()}`
+                            : `Created ${new Date(schedule.created_at).toLocaleDateString()}`}
+                        </p>
+                      </div>
+                      <div className="flex space-x-2">
+                        <button
+                          className="p-2 text-gray-400 hover:text-indigo-400"
+                          title="Download"
+                          onClick={(e) => { e.stopPropagation(); handleExportClick(schedule.id, "schedule"); }}
+                          disabled={exportingItemId === schedule.id}
+                        >
+                          {exportingItemId === schedule.id ? <Loader className="h-5 w-5 animate-spin" /> : <Download className="h-5 w-5" />}
+                        </button>
+                        <button
+                          className="p-2 text-gray-400 hover:text-indigo-400"
+                          title="Edit"
+                          onClick={() => handleEditProductionSchedule(schedule.id)}
+                        >
+                          <Edit className="h-5 w-5" />
+                        </button>
+                        <button
+                          className="p-2 text-gray-400 hover:text-red-400"
+                          title="Delete"
+                          onClick={() => handleDeleteRequest(schedule.id, "schedule", schedule.name)}
+                        >
+                          <Trash2 className="h-5 w-5" />
+                        </button>
+                      </div>
+                    </div>
+                  ))}
+                </div>
+              ) : (
+                <div className="p-4 bg-gray-700 rounded-lg text-center">
+                  <p className="text-gray-300 mb-4">You haven't created any production schedules yet</p>
+                </div>
+              )}
+              <div className="pt-3 text-center">
+                <div className="flex flex-col space-y-3 sm:flex-row sm:space-y-0 sm:space-x-3 sm:justify-center">
+                  <button
+                    className="inline-flex items-center bg-indigo-600 hover:bg-indigo-700 text-white px-4 py-2 rounded-md font-medium transition-all duration-200"
+                    onClick={handleCreateProductionSchedule}
+                  >
+                    <PlusCircle className="h-5 w-5 mr-2" />
+                    New Production Schedule
+                  </button>
+                  {productionSchedules.length > 0 && (
+                    <button
+                      className="inline-flex items-center bg-gray-700 hover:bg-gray-600 text-white px-4 py-2 rounded-md font-medium transition-all duration-200"
+                      onClick={() => navigate("/all-production-schedules")}
+                    >
+                      <List className="h-5 w-5 mr-2" />
+                      View All {productionSchedules.length > 0 && `(${productionSchedules.length})`}
+                    </button>
+                  )}
+                </div>
+              </div>
+            </div>
+          </div>
+
+          {/* My Run of Shows Card */}
+          <div className="bg-gray-800 p-6 rounded-xl shadow-md hover:shadow-lg transition-all duration-200">
+            <div className="flex justify-between items-start mb-6">
+              <div>
+                <h2 className="text-xl font-semibold text-white mb-2">My Run of Shows</h2>
+                <p className="text-gray-400">Organize and detail your show sequences</p>
+              </div>
+              <ClipboardList className="h-8 w-8 text-indigo-400" />
+            </div>
+            <div className="space-y-4">
+              {runOfShows.length > 0 ? (
+                <div className="space-y-3">
+                  {runOfShows.slice(0, 3).map((ros) => (
+                    <div
+                      key={ros.id}
+                      className="bg-gray-700 p-4 rounded-lg flex justify-between items-center"
+                    >
+                      <div>
+                        <h3 className="text-white font-medium">{ros.name}</h3>
+                        <p className="text-gray-400 text-sm">
+                          {ros.last_edited
+                            ? `Edited ${new Date(ros.last_edited).toLocaleDateString()}`
+                            : `Created ${new Date(ros.created_at).toLocaleDateString()}`}
+                        </p>
+                      </div>
+                      <div className="flex space-x-2">
+                        <button
+                          className="p-2 text-gray-400 hover:text-indigo-400"
+                          title="Download"
+                          onClick={(e) => { e.stopPropagation(); handleExportClick(ros.id, "runofshow"); }}
+                          disabled={exportingItemId === ros.id}
+                        >
+                          {exportingItemId === ros.id ? <Loader className="h-5 w-5 animate-spin" /> : <Download className="h-5 w-5" />}
+                        </button>
+                        <button
+                          className="p-2 text-gray-400 hover:text-indigo-400"
+                          title="Edit"
+                          onClick={() => handleEditRunOfShow(ros.id)}
+                        >
+                          <Edit className="h-5 w-5" />
+                        </button>
+                        <button
+                          className="p-2 text-gray-400 hover:text-red-400"
+                          title="Delete"
+                          onClick={() => handleDeleteRequest(ros.id, "runofshow", ros.name)}
+                        >
+                          <Trash2 className="h-5 w-5" />
+                        </button>
+                      </div>
+                    </div>
+                  ))}
+                </div>
+              ) : (
+                <div className="p-4 bg-gray-700 rounded-lg text-center">
+                  <p className="text-gray-300 mb-4">You haven't created any run of shows yet</p>
+                </div>
+              )}
+              <div className="pt-3 text-center">
+                <div className="flex flex-col space-y-3 sm:flex-row sm:space-y-0 sm:space-x-3 sm:justify-center">
+                  <button
+                    className="inline-flex items-center bg-indigo-600 hover:bg-indigo-700 text-white px-4 py-2 rounded-md font-medium transition-all duration-200"
+                    onClick={handleCreateRunOfShow}
+                  >
+                    <PlusCircle className="h-5 w-5 mr-2" />
+                    New Run of Show
+                  </button>
+                  {runOfShows.length > 0 && (
+                    <button
+                      className="inline-flex items-center bg-gray-700 hover:bg-gray-600 text-white px-4 py-2 rounded-md font-medium transition-all duration-200"
+                      onClick={() => navigate("/all-run-of-shows")}
+                    >
+                      <List className="h-5 w-5 mr-2" />
+                      View All {runOfShows.length > 0 && `(${runOfShows.length})`}
+                    </button>
+                  )}
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+      </main>
+
+      <ExportModal
+        isOpen={showProductionScheduleExportModal}
+        onClose={() => {if(!(exportingItemId && exportProductionScheduleId)) setShowProductionScheduleExportModal(false);}}
+        onExportImage={() => exportProductionScheduleId && prepareAndExecuteProductionScheduleExport(exportProductionScheduleId, 'image')}
+        onExportPdf={() => exportProductionScheduleId && prepareAndExecuteProductionScheduleExport(exportProductionScheduleId, 'print')}
+        title="Production Schedule"
+        isExporting={!!(exportingItemId && exportProductionScheduleId)}
+      />
+
+      <ExportModal
+        isOpen={showRunOfShowExportModal}
+        onClose={() => { if(!(exportingItemId && exportRunOfShowId)) setShowRunOfShowExportModal(false);}}
+        onExportImage={() => exportRunOfShowId && prepareAndExecuteRunOfShowExport(exportRunOfShowId, 'image')}
+        onExportPdf={() => exportRunOfShowId && prepareAndExecuteRunOfShowExport(exportRunOfShowId, 'print')}
+        title="Run of Show"
+        isExporting={!!(exportingItemId && exportRunOfShowId)}
+      />
+
+      {currentExportProductionSchedule && (
+        <>
+          <ProductionScheduleExport
+            key={`export-ps-${currentExportProductionSchedule.id}-${currentExportProductionSchedule.last_edited || currentExportProductionSchedule.created_at}`}
+            ref={productionScheduleExportRef}
+            schedule={currentExportProductionSchedule}
+          />
+          <PrintProductionScheduleExport
+            key={`print-export-ps-${currentExportProductionSchedule.id}-${currentExportProductionSchedule.last_edited || currentExportProductionSchedule.created_at}`}
+            ref={printProductionScheduleExportRef}
+            schedule={currentExportProductionSchedule}
+          />
+        </>
+      )}
+
+      {currentExportRunOfShow && (
+        <>
+          <RunOfShowExport
+            key={`export-ros-${currentExportRunOfShow.id}-${currentExportRunOfShow.last_edited || currentExportRunOfShow.created_at}`}
+            ref={runOfShowExportRef}
+            schedule={currentExportRunOfShow}
+          />
+          <PrintRunOfShowExport
+            key={`print-export-ros-${currentExportRunOfShow.id}-${currentExportRunOfShow.last_edited || currentExportRunOfShow.created_at}`}
+            ref={printRunOfShowExportRef}
+            schedule={currentExportRunOfShow}
+          />
+        </>
+      )}
+
+      {showDeleteConfirm && documentToDelete && (
+        <div className="fixed inset-0 bg-black/70 flex items-center justify-center z-50 p-4">
+          <div className="bg-gray-800 rounded-xl p-6 w-full max-w-md shadow-xl">
+            <div className="flex items-start">
+              <div className="flex-shrink-0 flex items-center justify-center h-12 w-12 rounded-full bg-red-500/20 sm:mx-0 sm:h-10 sm:w-10">
+                <AlertTriangle className="h-6 w-6 text-red-400" aria-hidden="true" />
+              </div>
+              <div className="ml-4 text-left">
+                <h3 className="text-lg font-medium text-white" id="modal-title">
+                  Delete {documentToDelete.type === "schedule" ? "Production Schedule" : "Run of Show"}
+                </h3>
+                <div className="mt-2">
+                  <p className="text-sm text-gray-300">
+                    Are you sure you want to delete "{documentToDelete.name}"? This action cannot be undone.
+                  </p>
+                </div>
+              </div>
+            </div>
+            <div className="mt-6 sm:mt-8 sm:flex sm:flex-row-reverse">
+              <button
+                type="button"
+                className="w-full inline-flex justify-center rounded-md border border-transparent shadow-sm px-4 py-2 bg-red-600 text-base font-medium text-white hover:bg-red-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-red-500 sm:ml-3 sm:w-auto sm:text-sm transition-colors"
+                onClick={confirmDelete}
+              >
+                Delete
+              </button>
+              <button
+                type="button"
+                className="mt-3 w-full inline-flex justify-center rounded-md border border-gray-600 shadow-sm px-4 py-2 bg-gray-700 text-base font-medium text-gray-300 hover:bg-gray-600 hover:text-white focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500 sm:mt-0 sm:w-auto sm:text-sm transition-colors"
+                onClick={cancelDelete}
+              >
+                Cancel
+              </button>
+            </div>
+          </div>
+        </div>
+      )}
+      <Footer />
+    </div>
+  );
+};
+
+export default ProductionPage;
